Id is now a Long, instead of a long

+++ src/test/java/com/jomm/terroir/business/AbstractUserTest.java
@@ -34,7 +34,7 @@ public class AbstractUserTest {
 	 */
 	@Test
 	public final void testGettersSetters() {
-		long nb = 0;
+		Long nb = (long) 0;
 		String test = "test";
 		
 		// Id
